requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coverage==5.5
cryptography==2.7
Django==2.2.5
django-appconf==1.0.3
django-cleanup==4.0.0
django-cors-headers==3.1.0
django-crispy-forms==1.11.1
django-imagekit==4.0.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
gunicorn==19.9.0
idna==2.8
numpy==1.17.2
panda==0.3.1
pandas==0.25.1
paramiko==2.6.0
pilkit==2.0
Pillow==6.1.0
psycopg2-binary==2.8.6
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pysftp==0.2.9
python-dateutil==2.8.0
python-slugify==4.0.0
pytz==2019.2
requests==2.22.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.5
wincertstore==0.2
